Rename scene management namespace to scene
