+++ Assets/Scripts/Architecture/Bootstrap/SceneBootstrapper.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 using WizardSpells.Data.Static.Enumerations.Scene;
-using WizardSpells.Services.SceneManagement;
+using WizardSpells.Services.Scene;
 using WizardSpells.Utilities.Patterns.Strategy;
 using Zenject;
