+++ Assets/Tests/EditMode/TestUtilities/Create.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 using WizardSpells.Architecture.GameStates.InitialSettingStrategies;
 using WizardSpells.Data.Static.Enumerations.Scene;
 using WizardSpells.Infrastructure.Factories.Components;
-using WizardSpells.Services.SceneManagement;
+using WizardSpells.Services.Scene;
 using WizardSpells.Utilities.Patterns.State.Machines;
 using WizardSpells.Utilities.Patterns.Strategy;
 using Zenject;
